requirements.txt
--trusted-host pypi.douban.com
--index-url http://pypi.douban.com/simple/

gevent
termcolor
pika
msgpack-python
PyYAML